:recycle: Move common test constants to p2p_test.go file
